+++ packages/widget-playground/src/defaultWidgetConfig.ts
@@ -41,11 +41,11 @@ export const widgetBaseConfig: WidgetConfig = {
   ],
   variant: 'wide',
   // subvariant: 'split',
-  subvariantOptions: {
-    wide: {
-      enableChainSidebar: true,
-    },
-  },
+  // subvariantOptions: {
+  //   wide: {
+  //     disableChainSidebar: true,
+  //   },
+  // },
   integrator: 'li.fi-playground',
   // fee: 0.01,
   // feeConfig: {
@@ -71,10 +71,14 @@ export const widgetBaseConfig: WidgetConfig = {
     apiUrl: 'https://li.quest/v1',
     chainsRefetchInterval: 1000 * 60 * 5, // 5 minutes
     rpcUrls: {
-      [ChainId.SOL]: ['https://dacey-pp61jd-fast-mainnet.helius-rpc.com/'],
+      [ChainId.SOL]: [
+        'https://wild-winter-frog.solana-mainnet.quiknode.pro/2370a45ff891f6dc9e5b1753460290fe0f1ef103/',
+        'https://dacey-pp61jd-fast-mainnet.helius-rpc.com/',
+      ],
     },
     routeOptions: {
       maxPriceImpact: 0.4,
+      jitoBundle: true,
       // slippage: 0.03,
       // order: 'SAFEST',
       // allowSwitchChain: false,